Add options for --from-auth-user and --to-auth-user
This commit adds options for specifying a user to authenticate as when
using the PLAIN mechanism. This is so we can migrate everyone's existing
email accounts without having to reset passwords, e.g:

  larch \
    --database ~/.larch/joe.bloggs.db \
    --sync-flags --all \
    --from imaps://mailhost.unboxedconsulting.com \
    --from-auth-user "joe.bloggs" \
    --from-user "admin" \
    --from-pass "password" \
    --to imaps://imap.gmail.com \
    --to-user "joe.bloggs@unboxedconsulting.com" \
    --to-pass "password"

Since the implementation in larch uses URIs we have to use the fragment
to identify which user we want to authenticate for which seems a little
hacky but it works well in testing.