Map Gmail folders from existing folders
This commit adds some special handling of folders for Gmail so that
the various types of deleted message folders, spam folders and sent
folders are amalgamated during migration.

The special handling is necessary because Gmail IMAP folders are
actually labels at there is special labels that begin with [Gmail]
for the system folders like 'Sent' and 'Trash'.

Two special cases are handled in the exception block:

1.  Where a user has selected UK language in the Gmail settings then
    the Trash folder will be called Bin and we return that instead.

2.  The [Gmail]/Chats folder is not visible by default in the list
    of IMAP folders so we need to raise an error for those cases.